LayoutTests/imported/w3c/web-platform-tests/custom-elements/form-associated/upgrading/constructor-form.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Custom elements don't get associated with a form in precustomized state</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<form id="theForm">
    <x-foo></x-foo>
</form>

<script>
test(() => {
    const callbackHistory = [];

    customElements.define("x-foo", class extends HTMLElement {
        static formAssociated = true;

        constructor() {
            super();

            window.internals = this.attachInternals();
            window.formInConstructor = internals.form;
        }

        formAssociatedCallback(form) {
            callbackHistory.push(`formAssociatedCallback: ${form.id}`);
        }
    });

    assert_equals(formInConstructor, null);
    assert_equals(internals.form, theForm);
    assert_array_equals(callbackHistory, ["formAssociatedCallback: theForm"]);
});
</script>
